Fix values for directions
The values for the directions as inputs into the intcode program were
incorrect. Specifically, South and East were swapped. My natural order
is North, East, South, and West. I initially put the values
correspondingly as 1, 2, 3, and 4, respectively. However, the actual
values (as specified by the description) are 1, 3, 2, and 4. The
description lists them as North, South, East, and West.